+++ test/test_faiss_and_milvus.py
@@ -13,6 +13,8 @@ from haystack.document_stores.weaviate import WeaviateDocumentStore
 from haystack.pipelines import Pipeline
 from haystack.nodes.retriever.dense import EmbeddingRetriever
 
+from conftest import ensure_ids_are_correct_uuids
+
 
 DOCUMENTS = [
     {"meta": {"name": "name_1", "year": "2020", "month": "01"}, "content": "text_1", "embedding": np.random.rand(768).astype(np.float32)},
@@ -445,3 +447,22 @@ def test_faiss_passing_index_from_outside(tmp_path):
     # test if vectors ids are associated with docs
     for doc in documents_indexed:
         assert 0 <= int(doc.meta["vector_id"]) <= 7
+
+
+@pytest.mark.parametrize("document_store_small", ["faiss", "milvus", "weaviate"], indirect=True)
+def test_cosine_sanity_check(document_store_small):
+    VEC_1 = np.array([.1, .2, .3], dtype="float32")
+    VEC_2 = np.array([.4, .5, .6], dtype="float32")
+
+    # This is the cosine similarity of VEC_1 and VEC_2 calculated using sklearn.metrics.pairwise.cosine_similarity
+    # The score is normalized to yield a value between 0 and 1.
+    KNOWN_COSINE = (0.9746317 + 1) / 2
+
+    docs = [{"name": "vec_1", "text": "vec_1", "content": "vec_1", "embedding": VEC_1}]
+    ensure_ids_are_correct_uuids(docs=docs,document_store=document_store_small)
+    document_store_small.write_documents(documents=docs)
+
+    query_results = document_store_small.query_by_embedding(query_emb=VEC_2, top_k=1, return_embedding=True)
+
+    # check if faiss returns the same cosine similarity. Manual testing with faiss yielded 0.9746318
+    assert math.isclose(query_results[0].score, KNOWN_COSINE, abs_tol=0.00002)
